refactor(users): move page query logic into feature hooks
